feat: 회원가입 페이지 캐싱 전략 및 메타데이터 추가, 닉네임 중복확인 API Route 이동
